Create packages by screen and move files
